requirements.txt
selenium==4.9.1
webdriver_manager==4.0.0
lxml
boto3
PyYAML
undetected-chromedriver